separate parser tests to integrations tests
